Refactor and add tests
Adding tests for the jenkins handler and refactoring. Removed dependency
on "SaladJenkinsTracking" which was not in the original gemfile. Updated
date rendering to use the same format as the chef-tracking Jenkins
plugin expects. Changed the way the md5 is calculated so that it reads
files 1 MB at a time, rather than in one go, which can improve
performance for large files. Using Net::HTTP for submission rather than
the Chef one, so that authorization headers are not transmitted to
Jenkins. Using the JSON gem to render the json. Added a 'dryrun'
parameter so that test data can be viewed but not necessarily submitted
to Jenkins.